Move code for writing CrystFEL geometry to separate module
